+++ src/Components/CodeLens.fs
@@ -12,85 +12,88 @@ open Ionide.VSCode.Helpers
 
 module CodeLens =
     let private createProvider () =
-        let mapRes (doc : TextDocument) o =
-             o.Data |> Array.collect (fun syms ->
-                let range = Range
-                                ( float syms.Declaration.BodyRange.StartLine - 1.,
-                                    float syms.Declaration.BodyRange.StartColumn - 1.,
-                                    float syms.Declaration.BodyRange.EndLine - 1.,
-                                    float syms.Declaration.BodyRange.EndColumn - 1.)
-                let cl = CodeLens(range)
+        let symbolsToCodeLens (doc : TextDocument) (symbols: Symbols[]) : CodeLens[] =
+             symbols |> Array.collect (fun syms ->
+                let range = CodeRange.fromDTO syms.Declaration.BodyRange
+                let codeLens = CodeLens range
 
-                let cls =  syms.Nested |> Array.choose (fun sym ->
-                    if sym.GlyphChar <> "Fc" && sym.GlyphChar <> "M" then None
-                    elif sym.Glyph = "Extension Method" then
-                        Range
-                            ( float sym.BodyRange.StartLine - 1.,
-                                float sym.BodyRange.StartColumn - (float sym.Name.Length),
-                                float sym.BodyRange.EndLine - 1.,
-                                float sym.BodyRange.EndColumn - 1.)
-                        |> CodeLens |> Some
-                    else
-                        Range
-                            ( float sym.BodyRange.StartLine - 1.,
-                                float sym.BodyRange.StartColumn - 1.,
-                                float sym.BodyRange.EndLine - 1.,
-                                float sym.BodyRange.EndColumn - 1.)
-                        |> CodeLens |> Some )
-                if syms.Declaration.GlyphChar <> "Fc" then cls
-                else
-                    cls |> Array.append (Array.create 1 cl))
+                let codeLenses =  syms.Nested |> Array.choose (fun sym ->
+                    if sym.GlyphChar <> "Fc"
+                       && sym.GlyphChar <> "M"
+                       && sym.GlyphChar <> "F"
+                       || sym.IsAbstract
+                       || sym.EnclosingEntity = "I"  // interface
+                       || sym.EnclosingEntity = "R"  // record
+                       || sym.EnclosingEntity = "D"  // DU
+                       || sym.EnclosingEntity = "En" // enum
+                       || sym.EnclosingEntity = "E"  // exception
+                    then None
+                    else Some (CodeLens (CodeRange.fromDTO sym.BodyRange)))
 
-        let toSingature (t : string) =
-            let t =
-                if t.StartsWith "val" || t.StartsWith "member" || t.StartsWith "abstract" then
-                    t
+                if syms.Declaration.GlyphChar <> "Fc" then codeLenses
                 else
-                    let i = t.IndexOf("(")
+                    codeLenses |> Array.append [|codeLens|])
+
+        let formatSingature (sign : string) : string =
+            let sign =
+                if sign.StartsWith "val" || sign.StartsWith "member" || sign.StartsWith "abstract" then
+                    sign
+                else
+                    let i = sign.IndexOf("(")
                     if i > 0 then
-                        t.Substring(0, i) + ":" + t.Substring(i+1)
+                        sign.Substring(0, i) + ":" + sign.Substring(i+1)
                     else
-                        t
+                        sign
 
-            let sign = if t.Contains(":") then t.Split(':').[1 ..] |> String.concat ":" else t
+            let sign = if sign.Contains(":") then sign.Split(':').[1 ..] |> String.concat ":" else sign
             let parms = sign.Split( [|"->"|], StringSplitOptions.RemoveEmptyEntries)
-            let nSing =
-                parms |> Seq.map (fun p ->
-                    if p.Contains "(requires" then
-                        p
-                    elif p.Contains "*" then
-                        p.Split('*') |> Seq.map (fun z -> if z.Contains ":" then z.Split(':').[1] else z) |> String.concat "* "
-                    elif p.Contains "," then
-                        p.Split(',') |> Seq.map (fun z -> if z.Contains ":" then z.Split(':').[1] else z) |> String.concat "* "
-                    elif p.Contains ":" then
-                        p.Split(':').[1]
-                    else p
-                ) |> Seq.map (fun p -> p.Trim()) |> String.concat " -> "
-            nSing.Replace("<", "&lt;").Replace(">", "&gt;")
+            parms
+            |> Seq.map (fun p ->
+                if p.Contains "(requires" then
+                    p
+                elif p.Contains "*" then
+                    p.Split '*' |> Seq.map (fun z -> if z.Contains ":" then z.Split(':').[1] else z) |> String.concat "* "
+                elif p.Contains "," then
+                    p.Split ',' |> Seq.map (fun z -> if z.Contains ":" then z.Split(':').[1] else z) |> String.concat "* "
+                elif p.Contains ":" then
+                    p.Split(':').[1]
+                else p)
+            |> Seq.map String.trim
+            |> String.concat " -> "
+            |> String.replace "<" "&lt;"
+            |> String.replace ">" "&gt;"
 
-
-        { new CodeLensProvider
-          with
-            member this.provideCodeLenses(doc, ct) =
+        { new CodeLensProvider with
+            member __.provideCodeLenses(doc, _) =
                 promise {
                     let! _ = LanguageService.parse doc.fileName (doc.getText())
-                    let! o = LanguageService.declarations doc.fileName
-                    let data = mapRes doc o
-                    Browser.console.log("Provide", data)
-                    return data |> ResizeArray
+                    let! result = LanguageService.declarations doc.fileName
+                    let data = symbolsToCodeLens doc result.Data
+                    return ResizeArray data
                 } |> Case2
 
 
-            member this.resolveCodeLens(cl, ct) =
+            member __.resolveCodeLens(codeLens, _) =
                 promise {
-                    let! o = LanguageService.toolbar (window.activeTextEditor.document.fileName) (int cl.range.start.line + 1) (int cl.range.start.character + 1)
-                    let res = (o.Data |> Array.fold (fun acc n -> (n |> Array.toList) @ acc ) []).Head.Signature
-                    let sign = res.Split('\n').[0].Trim() |> toSingature
+                    let! signaturesResult =
+                        LanguageService.toolbar
+                            window.activeTextEditor.document.fileName
+                            (int codeLens.range.start.line + 1)
+                            (int codeLens.range.start.character + 1)
+
+                    let res =
+                        signaturesResult.Data
+                        |> Array.rev
+                        |> Array.tryHead
+                        |> Option.bind Array.tryHead
+                        |> Option.map (fun x -> x.Signature)
+                        |> Option.fill ""
+
+                    let sign = res.Split('\n').[0] |> String.trim |> formatSingature
                     let cmd = createEmpty<Command>
                     cmd.title <- sprintf "%s" sign
-                    cl.command <- cmd
-                    Browser.console.log("Resolve", cl, res)
-                    return cl
+                    codeLens.command <- cmd
+                    return codeLens
                 } |> Case2
         }