+++ src/Components/Fsi.fs
@@ -11,14 +11,11 @@ open DTO
 open Ionide.VSCode.Helpers
 
 module Fsi =
-    [<Emit("setTimeout($0,$1)")>]
-    let setTimeout(cb, delay) : obj = failwith "JS Only"
-
     let mutable fsiOutput : Terminal option = None
     let mutable fsiOutputPID : int option = None
 
     let isPowershell () =
-        let t = workspace.getConfiguration().get("terminal.integrated.shell.windows", "")
+        let t = "terminal.integrated.shell.windows" |> Configuration.get ""
         t.ToLower().Contains "powershell"
 
     let sendCd () =
@@ -46,7 +43,9 @@ module Fsi =
             fsiOutput |> Option.iter (fun n -> n.dispose())
             let parms =
                 let fsiParams =
-                    workspace.getConfiguration().get("FSharp.fsiExtraParameters", Array.empty<string>) |> List.ofArray
+                    "FSharp.fsiExtraParameters"
+                    |> Configuration.get Array.empty<string>
+                    |> List.ofArray
 
                 if Environment.isWin then
                     [ "--fsi-server-input-codepage:65001" ] @ fsiParams