+++ src/Components/Rename.fs
@@ -14,7 +14,7 @@ module Rename =
     let private createProvider () =
         let mapResult (doc : TextDocument) (newName : string) (o : SymbolUseResult) =
             let res = WorkspaceEdit ()
-            if o |> unbox <> null then
+            if isNotNull o then
                 o.Data.Uses |> Array.iter (fun s ->
                     let range = Range(float s.StartLine - 1., (float s.EndColumn) - (float o.Data.Name.Length) - 1., float s.EndLine - 1., float s.EndColumn - 1.)
                     let te = TextEdit.replace(range, newName)