+++ src/Components/Symbols.fs
@@ -30,18 +30,14 @@ module Symbols =
             | _   -> 0 |> unbox
 
         let mapRes (doc : TextDocument) o =
-             if o |> unbox <> null then
+             if isNotNull o then
                 o.Data |> Array.map (fun syms ->
                     let oc = createEmpty<SymbolInformation>
                     oc.name <- syms.Declaration.Name
                     oc.kind <- syms.Declaration.GlyphChar |> convertToKind
                     oc.containerName <- syms.Declaration.Glyph
                     let loc = createEmpty<Location>
-                    loc.range <-  Range
-                                ( float syms.Declaration.BodyRange.StartLine   - 1.,
-                                    float syms.Declaration.BodyRange.StartColumn - 1.,
-                                    float syms.Declaration.BodyRange.EndLine     - 1.,
-                                    float syms.Declaration.BodyRange.EndColumn   - 1.)
+                    loc.range <- CodeRange.fromDTO syms.Declaration.BodyRange
                     loc.uri <- Uri.file doc.fileName
                     oc.location <- loc
                     let ocs =  syms.Nested |> Array.map (fun sym ->
@@ -50,11 +46,7 @@ module Symbols =
                         oc.kind <- sym.GlyphChar |> convertToKind
                         oc.containerName <- sym.Glyph
                         let loc = createEmpty<Location>
-                        loc.range <-  Range
-                                    ( float sym.BodyRange.StartLine   - 1.,
-                                        float sym.BodyRange.StartColumn - 1.,
-                                        float sym.BodyRange.EndLine     - 1.,
-                                        float sym.BodyRange.EndColumn   - 1.)
+                        loc.range <- CodeRange.fromDTO sym.BodyRange
                         loc.uri <- Uri.file doc.fileName
                         oc.location <- loc
                         oc )